+++ 02.Domain/02.A.DomainTests/DomainTest/UserAccounting/Roles/RoleTestBuilder.cs
@@ -9,7 +9,7 @@ namespace DomainTest.UserAccounting.Roles
     {
         private readonly RoleTestBuilder _builderInstance = null;
 
-        public int Id = someId;
+        public long Id = someId;
         public string Title = someTitle;
         public string SystemDescription = someSystemDescription;
         public string Description = someDescription;
