+++ 02.Domain/02.B.DomainModels/Domain/UserAccounting/Positions/PositionBuilder.cs
@@ -18,7 +18,7 @@ namespace Domain.UserAccounting.Positions
         public string CustomesCode = someCustomesCode;
         public string Description = someDescription;
         public bool IsActive = someIsActive;
-        public int RoleId = someRoleId;
+        public long RoleId = someRoleId;
 
         public PositionBuilder()
         {
