+++ 02.Domain/02.B.DomainModels/Domain/UserAccounting/Roles/Role.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+
 using DomainServiceContract.Roles;
 using DomainShared.BaseAggregates;
 using ExceptionsManagement.DomainExceptions.Roles;
@@ -8,17 +9,17 @@ using Utilities.SharedTools.ExceptionDictionaries;
 
 namespace Domain.UserAccounting.Roles
 {
-    public class Role:BaseDomainEntity<int>, IAggregateRoot
+    public class Role : BaseDomainEntity<long>, IAggregateRoot
     {
         public string Title { get; private set; }
         public string SystemDescription { get; private set; }
-        public string Description { get;private set; }
+        public string Description { get; private set; }
 
         private IDomainRoleExceptionHelper _roleExceptionHelper = new RoleExceptionHelper();
 
         public event EventHandler<RoleEventArgs> RoleCreated;
 
-        public Role(int id, string title, string systemDescription,string description):base(id)
+        public Role(long id, string title, string systemDescription, string description) : base(id)
         {
 
             if (title.Length > 100 || string.IsNullOrWhiteSpace(title) || string.IsNullOrEmpty(title))
@@ -31,15 +32,15 @@ namespace Domain.UserAccounting.Roles
                 _roleExceptionHelper.ThrowExceptionMessage((long)ExceptionCodes.UserAccountingDomainRoleSystemDescription);
             }
 
-            if (description.Length>500 || string.IsNullOrWhiteSpace(description)|| string.IsNullOrEmpty(description))
+            if (description.Length > 500 || string.IsNullOrWhiteSpace(description) || string.IsNullOrEmpty(description))
             {
-                _roleExceptionHelper.ThrowExceptionMessage((long) ExceptionCodes.UserAccountingDomainRoleDescription);
+                _roleExceptionHelper.ThrowExceptionMessage((long)ExceptionCodes.UserAccountingDomainRoleDescription);
             }
 
             Title = title;
             SystemDescription = systemDescription;
             Description = description;
-            
+
             OnRoleCreated(this);
         }
 
@@ -47,7 +48,7 @@ namespace Domain.UserAccounting.Roles
         {
             if (RoleCreated != null)
             {
-                RoleCreated(this, new RoleEventArgs(){Role = role});
+                RoleCreated(this, new RoleEventArgs() { Role = role });
             }
         }
