+++ 07.UIAndSDKTest/B.SecondaryUI/TestBridge/Models/UserAccounting/Roles/BridgeRoleViewModel.cs
@@ -2,13 +2,13 @@
 
 namespace TestBridge.Models.UserAccounting.Roles
 {
-    public class BridgeRoleViewModel:BaseViewModel<int>
+    public class BridgeRoleViewModel : BaseViewModel<long>
     {
         public string Title { get; private set; }
         public string SystemDescription { get; private set; }
         public string Description { get; private set; }
 
-        public BridgeRoleViewModel(int id, string title, string systemDescription, string description) : base(id)
+        public BridgeRoleViewModel(long id, string title, string systemDescription, string description) : base(id)
         {
             Title = title;
             SystemDescription = systemDescription;
